Good news, everyone!
Fixed the algorithm, there was a flipped greater than operand that made everything fall apart. I threw some garbage at it and it seemed to handle it just fine. That being said, I have not had time to sanitize input (I didn't want to accidentally break Spring by returning an exception it wasn't expecting to handle).

I also added an improvement to the time generators for the instance of orders that take place at the same "time"; it adds the average to parcel processed last.

The averaging had to be done in MinToHours() or it would mess with the 2400 hour conversion.